scraper/requirements.txt
feedparser
supabase
python-dotenv
google-generativeai
cloudscraper
beautifulsoup4
libsql-client
libsql